Move index file to index.js
